refactor(general_functions_source): rename modul to `utils`
Rename module `general_functions_source` to `utils`.
The term utils is more widely known and used for modules, which
are collections of semantically not coherent, and widely
applicable methods.
