+++ invenio_rdm_pure/cli.py
@@ -29,7 +29,7 @@ from docopt import docopt
 
 from .setup import dirpath
 from .shell_interface import ShellInterface, method_call
-from .source.general_functions_source import check_if_directory_exists
+from .source.utils import check_if_directory_exists
 
 if __name__ == "__main__":
     arguments = docopt(__doc__, version="Pure synchronizer 1.0")
