+++ invenio_rdm_pure/source/rdm/delete_record.py
@@ -8,8 +8,8 @@
 """File description."""
 
 from ...setup import data_files_name, push_dist_sec
-from ..general_functions_source import current_time, file_read_lines
 from ..reports import Reports
+from ..utils import current_time, file_read_lines
 from .requests_rdm import Requests
