+++ invenio_rdm_pure/source/rdm/run/owners.py
@@ -11,14 +11,9 @@ import json
 from datetime import date, datetime
 
 from ....setup import data_files_name, pure_uuid_length
-from ...general_functions_source import (
-    add_spaces,
-    file_read_lines,
-    initialize_counters,
-    shorten_file_name,
-)
 from ...pure.requests_pure import get_next_page, get_pure_metadata
 from ...reports import Reports
+from ...utils import add_spaces, file_read_lines, initialize_counters, shorten_file_name
 from ..add_record import RdmAddRecord
 from ..database import RdmDatabase
 from ..record_manager import RecordManager
