+++ invenio_rdm_pure/source/rdm/versioning.py
@@ -9,8 +9,8 @@
 
 import json
 
-from ..general_functions_source import add_spaces
 from ..reports import Reports
+from ..utils import add_spaces
 from .record_manager import RecordManager
 from .requests_rdm import Requests
